chatbot/requirements.txt
Flask
pandas
scikit-learn
numpy
gunicorn
tensorflow-cpu
matplotlib
seaborn
nltk